repro/requirements.txt
tensorflow==2.18.0
pandas
matplotlib
tensorboard-plugin-profile
# extra tensorboard-plugin-profile deps
etils
importlib-resources
keras_tuner
